pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "podly-pure-podcasts"
version = "0.0.0"
description = "Ad-block for podcasts. Create an ad-free RSS feed."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "openai",
    "python-dotenv",
    "jinja2",
    "flask",
    "pyrss2gen",
    "feedparser",
    "certifi",
    "cd",
    "pyyaml",
    "pypodcastparser",
    "werkzeug",
    "exceptiongroup",
    "waitress",
    "validators",
    "beartype",
    "flask-sqlalchemy",
    "flask-migrate",
    "Flask-APScheduler",
    "ffmpeg-python",
    "litellm",
    "bleach",
    "async_timeout",
    "flask-cors",
    "bcrypt",
    "httpx-aiohttp",
    "stripe",
    "mutagen",
    "github-copilot-sdk",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "ty",
    "types-bleach",
    "types-pyyaml",
    "types-requests",
    "types-waitress",
    "pytest",
    "pytest-cov",
    "dill",
    "types-flask-migrate",
    "pytest-mock",
    "watchdog",
    "requests",
    "types-flask-cors",
]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
norecursedirs = ["scripts", "scripts/*", ".worktrees", ".worktrees/*"]
testpaths = []

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "BLE", "SIM", "RUF", "PLC", "PLE", "PLR", "PLW", "ANN"]
ignore = [
    "E501",    # line-too-long (formatter handles it)
    # ANN — relaxed for gradual adoption
    "ANN001",  # missing-type-function-argument
    "ANN002",  # missing-type-args
    "ANN003",  # missing-type-kwargs
    "ANN201",  # missing-return-type-public-function
    "ANN202",  # missing-return-type-private-function
    "ANN204",  # missing-return-type-special-method
    "ANN206",  # missing-return-type-class-method
    "ANN401",  # dynamically-typed-expressions (Any)
    # PLC
    "PLC0415", # import-outside-toplevel (intentional lazy imports)
    # PLR — rules that were globally disabled in the old pylint config
    "PLR0911", # too-many-return-statements
    "PLR0912", # too-many-branches (intentional in complex processors/routes)
    "PLR0913", # too-many-arguments
    "PLR0914", # too-many-locals
    "PLR0915", # too-many-statements
    "PLR0917", # too-many-positional-arguments
    "PLR2004", # magic-value-comparison
    # PLW
    "PLW0603", # global-statement (intentional module-level singletons)
    "PLW1508", # invalid-envvar-default (false positive on os.environ.get)
    # RUF
    "RUF001",  # ambiguous-unicode-character-string (false positives on en-dash)
    "RUF012",  # mutable-class-variable (config classes)
    # SIM — style preferences
    "SIM102",  # collapsible-if
    "SIM103",  # needless-bool
    "SIM105",  # use-contextlib-suppress
    "SIM108",  # if-else-block-instead-of-if-exp
    "SIM117",  # multiple-with-statements
    "SIM118",  # in-dict-keys
    "SIM201",  # negate-equal-op
    # BLE — broad exception catches are intentional throughout (previously pylint: disable=broad-except)
    "BLE001",  # blind-exception
]

[tool.ruff.lint.per-file-ignores]
"src/tests/**/*.py" = ["ANN"]
"scripts/**/*.py" = ["ANN"]

[tool.ty.environment]
python-version = "3.11"

[tool.ty.src]
include = ["src"]

[tool.ty.rules]
unresolved-attribute = "warn"
invalid-argument-type = "warn"
invalid-return-type = "warn"
invalid-method-override = "warn"
invalid-assignment = "warn"
no-matching-overload = "warn"
not-subscriptable = "warn"
# deprecated: utcnow() is pervasive throughout the codebase — suppress until a
# dedicated timezone-aware migration is done.
deprecated = "ignore"
# unused-type-ignore-comment: suppress until mypy-targeted type: ignore comments
# are audited against ty.
unused-type-ignore-comment = "ignore"
